refactor(App, Lotto, LottoMachine):  LottoMachine 클래스로 관심사 분리 및 코드 이동
- generateLottos 함수의 반복문을 Array.from() 및 map으로 대체하여 가독성 향상
- App 클래스의 generateLottos 함수를 LottoMachine 클래스로 분리
- Lotto 클래스의 pickRandomNumbers 함수 LottoMachine 클래스로 분리
- LottoMachine 테스트 코드 분리
